requirements.txt
setuptools>=42
wheel

requests
python-dotenv